requirements-d-reduction.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements-d-reduction.in
#
joblib==0.13.2            # via scikit-learn
llvmlite==0.28.0          # via numba
numba==0.43.1             # via umap-learn
numpy==1.16.3
pandas==0.24.2
python-dateutil==2.8.0    # via pandas
pytz==2019.1              # via pandas
scikit-learn==0.21.0
scipy==1.2.1              # via scikit-learn, umap-learn
six==1.12.0               # via python-dateutil
umap-learn==0.3.8
